Refactor to create discrete auth service
